Added delete functionality with react-native-swipeout
Not sure that I like this library. It's not able to handle row
reuse in a FlatList. I'm probably going to try Wix's swipeout
library. If I don't like that I'll move on to a long press
to change the future "done" checkbox into a check circle to delete,
similar to Google Inbox or every single iOS app ever